feat(effects): implement effects system and UI components for Phase 3
- Add effects module with cursor tracking, zoom, keyboard capture, background compositing
- Add project format v1 with binary .frame files and effects config persistence
- Add keyboard badge widget for shortcut overlays during recording
- Add settings panel UI for configuring zoom, keyboard, and background effects
- Fix clippy warnings across workspace (derive patterns, lifetimes, dead code)

+++ packages/core/src/capture/platform.rs
@@ -64,8 +64,8 @@ pub mod macos {
                 .iter()
                 .map(|d| DisplayInfo {
                     id: d.display_id() as u64,
-                    width: d.width() as u32,
-                    height: d.height() as u32,
+                    width: d.width(),
+                    height: d.height(),
                     frame_rate: 60,
                 })
                 .collect();
@@ -151,7 +151,7 @@ pub mod macos {
                         .with_display(display)
                         .with_excluding_windows(&[])
                         .build();
-                    (filter, display.width() as u32, display.height() as u32)
+                    (filter, display.width(), display.height())
                 }
                 CaptureArea::Window { window_id } => {
                     let windows = content.windows();
